Manually require and export all util scripts
* Fixes #466
