+++ lib/utils/index.js
@@ -1,19 +1,28 @@
 'use strict'
 
-const { readdirSync } = require('fs')
-const { basename } = require('path')
-
-const files = readdirSync(__dirname).filter(f => f.endsWith('.test.js') === false)
-for (const file of files) {
-  if (file === 'index.js') continue
-
-  const kebabName = basename(file, '.js')
-  const snakeName = kebabName.split('-').map((part, idx) => {
-    if (idx === 0) return part
-    return part[0].toUpperCase() + part.slice(1)
-  }).join('')
-
-  module.exports[snakeName] = require(`./${kebabName}.js`)
+module.exports = {
+  buildSafeSonicBoom: require('./build-safe-sonic-boom.js'),
+  createDate: require('./create-date.js'),
+  deleteLogProperty: require('./delete-log-property.js'),
+  filterLog: require('./filter-log.js'),
+  formatTime: require('./format-time.js'),
+  getPropertyValue: require('./get-property-value.js'),
+  handleCustomLevelsNamesOpts: require('./handle-custom-levels-names-opts.js'),
+  handleCustomLevelsOpts: require('./handle-custom-levels-opts.js'),
+  interpretConditionals: require('./interpret-conditionals.js'),
+  isObject: require('./is-object.js'),
+  isValidDate: require('./is-valid-date.js'),
+  joinLinesWithIndentation: require('./join-lines-with-indentation.js'),
+  noop: require('./noop.js'),
+  parseFactoryOptions: require('./parse-factory-options.js'),
+  prettifyErrorLog: require('./prettify-error-log.js'),
+  prettifyError: require('./prettify-error.js'),
+  prettifyLevel: require('./prettify-level.js'),
+  prettifyMessage: require('./prettify-message.js'),
+  prettifyMetadata: require('./prettify-metadata.js'),
+  prettifyObject: require('./prettify-object.js'),
+  prettifyTime: require('./prettify-time.js'),
+  splitPropertyKey: require('./split-property-key.js')
 }
 
 // The remainder of this file consists of jsdoc blocks that are difficult to
